Add Arduino code for IMU Magnetometer
In the past the heading information for navigation was acquired from
midg which also provided gps information on Joemegatron.  This sensor
is not on Enterprise and has been replaced by a sensor stick from
sparkfun: https://www.sparkfun.com/products/10724.

This code only takes advantage of the magnetometer and also changes
the expected message of position to a single quaternion since the
IMU message is too large for the arduino to publish.

Type: New Feature