Add comments to find_twist()
find_twist() is the most important function in the pot_nav algorithm.
This funciton finds the linear and angular velocity to send to the
robot and ensures that it does not run in to any objects.

The function works by tracing rays on the image recieved by color stat.
When it finds an object it adjusts its course based on its distance to
that object (as it gets closer it slows and turns away).  After all
traces are finished it returns a twist that is based on the distance of
each object it has found.