Add new Manual Differential Drive mode to the control node.
The new mode is triggered using the X button on the gamepad, and uses
the left and right trigger keys to make the robot go forward and
backward, respectively. This will be changed in the future to control
the left and right robot wheels instead, but that requires changes to
the microcontroller code.