Add enterprise_description package
In order to use some of the more advanced ros features it is important
that enterprise have a transform tree which describes it.  The tf
package documentation can be found at: http://wiki.ros.org/tf.

This commit includes a node which takes data from the imu and transmits
changes in orientation to the odom frame.  This should be the first
step to implementing gmapping/SLAM.  It will be important to combine
this rotation data with translation data for SLAM to do more than map
in a circle.