Add a way to read waypoints from a file
Its easier to input waypoints to a file and import that file than it is
to use the parameter server to store the waypoints.  The file is
currently set to read from a file that contains latitudes and
longitudes seperated by a space.  It is expected that there is only one
waypoint per line.  It uses a queue to store the waypoints and only the
first is used at this time.