Add skeleton for hard stopping
The robot is going to need to be able to hardstop, this skeleton adds
a place to add the stop.  A hardware stop could be implemented by
adding an arduino message that goes to a relay  but right now software
stop is implemented where all components are told to stop moving if "b"
is pressed.